Updating supported swift versions list
This PR adds Swift 4.1.0 and 4.2.0 as supported versions in the docs. 4.1.0 is supported, for example it is used in [Vapor TIL app](https://github.com/raywenderlich/vapor-til/blob/master/cloud.yml), while I know for sure that also 4.2.0 is supported since I'm using it. If you think there are other versions in the middle we should add please comments are appreciated 😊